requirements.txt
# LED Matrix Control
rpi-rgb-led-matrix>=1.0.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# GPIO Access
RPi.GPIO>=0.7.1

# Spotify API
spotipy>=2.23.0

# HTTP Requests
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv>=1.0.0

# Stock Ticker
yfinance>=0.2.0
